requirements.txt
fastapi
uvicorn[standard]
pandas
numpy
scikit-learn
xgboost
pyyaml
python-dotenv
azure-storage-blob
wandb

# Dev
black
flake8
isort
